+++ src/by/it/novik/JD02_09/Automatic_Generation/scheme1.xsd
